# social-mini-app/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Social Intelligence Platform API",
    "description": "Public REST API for the Social Intelligence Platform. All endpoints require a valid API key. Internal keys have higher rate limits and (in Sprint 2) access to write endpoints; external keys are read-only and rate-limited.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/subjects/{subject_id}/alerts": {
      "get": {
        "tags": [
          "alerts"
        ],
        "summary": "List alert rules for a subject",
        "description": "Return all alert rules for a subject, newest first. Supports `active_only` filter and pagination.",
        "operationId": "list_alerts_v1_subjects__subject_id__alerts_get",
        "parameters": [
          {
            "name": "subject_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Subject Id"
            }
          },
          {
            "name": "active_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Only return active rules.",
              "default": false,
              "title": "Active Only"
            },
            "description": "Only return active rules."
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "1-indexed page number.",
              "default": 1,
              "title": "Page"
            },
            "description": "1-indexed page number."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page (max 100).",
              "default": 20,
              "title": "Limit"
            },
            "description": "Items per page (max 100)."
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope_list_AlertRule__"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key"
          },
          "404": {
            "description": "Subject not found"
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "alerts"
        ],
        "summary": "Create an alert rule for a subject",
        "description": "Create a new alert rule. Requires an internal API key.",
        "operationId": "create_alert_v1_subjects__subject_id__alerts_post",
        "parameters": [
          {
            "name": "subject_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Subject Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlertRuleCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope_AlertRule_"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key"
          },
          "403": {
            "description": "Internal API key required"
          },
          "404": {
            "description": "Subject not found"
          },
          "422": {
            "description": "Invalid request body"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        }
      }
    },
    "/v1/alerts/{rule_id}": {
      "put": {
        "tags": [
          "alerts"
        ],
        "summary": "Update an alert rule",
        "description": "Update an existing alert rule. Requires an internal API key.",
        "operationId": "update_alert_v1_alerts__rule_id__put",
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Rule Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlertRuleUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEnvelope_AlertRule_"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key"
          },
          "403": {
            "description": "Internal API key required"
          },
          "404": {
            "description": "Rule not found"
          },
          "422": {
            "description": "Invalid request body"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        }
      },
      "delete": {
        "tags": [
          "alerts"
        ],
        "summary": "Delete an alert rule",
        "description": "Delete an alert rule. Requires an internal API key.",
        "operationId": "delete_alert_v1_alerts__rule_id__delete",
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Rule Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Missing or invalid API key"
          },
          "403": {
            "description": "Internal API key required"
          },
          "404": {
            "description": "Rule not found"
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AlertRule": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "subject_id": {
            "type": "string",
            "format": "uuid",
            "title": "Subject Id"
          },
          "rule_type": {
            "$ref": "#/components/schemas/AlertRuleType"
          },
          "threshold": {
            "type": "number",
            "title": "Threshold"
          },
          "cooldown_seconds": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Cooldown Seconds"
          },
          "channel_id": {
            "type": "string",
            "title": "Channel Id"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "subject_id",
          "rule_type",
          "threshold",
          "cooldown_seconds",
          "channel_id"
        ],
        "title": "AlertRule",
        "description": "Configurable alert condition scoped to one subject.\n\nCreated and updated by analysts through the Mini App; evaluated by\nthe alert engine on each sync cycle. Phase 1 defines the contract\nand table; the engine itself ships in Phase 4."
      },
      "AlertRuleCreate": {
        "properties": {
          "rule_type": {
            "$ref": "#/components/schemas/AlertRuleType"
          },
          "threshold": {
            "type": "number",
            "minimum": 0.0,
            "title": "Threshold"
          },
          "cooldown_seconds": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Cooldown Seconds",
            "default": 3600
          },
          "channel_id": {
            "type": "string",
            "minLength": 1,
            "title": "Channel Id",
            "default": "channel_id"
          }
        },
        "type": "object",
        "required": [
          "rule_type",
          "threshold"
        ],
        "title": "AlertRuleCreate"
      },
      "AlertRuleType": {
        "type": "string",
        "enum": [
          "follower_spike",
          "follower_drop",
          "activity_spike",
          "activity_silence",
          "status_change"
        ],
        "title": "AlertRuleType",
        "description": "Type of alert condition evaluated by the alert engine."
      },
      "AlertRuleUpdate": {
        "properties": {
          "rule_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AlertRuleType"
              },
              {
                "type": "null"
              }
            ]
          },
          "threshold": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Threshold"
          },
          "cooldown_seconds": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Cooldown Seconds"
          },
          "channel_id": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel Id"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "AlertRuleUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResponseEnvelope_AlertRule_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AlertRule"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ResponseEnvelope[AlertRule]"
      },
...
      "ResponseEnvelope_list_AlertRule__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/AlertRule"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ResponseEnvelope[list[AlertRule]]"
      },
...
      "ResponseMeta": {
        "properties": {
          "page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          },
          "limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "type": "object",
        "title": "ResponseMeta",
        "description": "Standard response metadata.\n\nAlways present in success responses. Fields are optional so the\nsame shape works for single-resource and paginated list responses."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Admin token from ADMIN_TOKEN env var.",
        "scheme": "bearer"
      },
      "APIKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key issued via the admin endpoint. Internal keys have higher rate limits; external keys are read-only. The raw key is shown exactly once at creation time."
      },
      "AdminAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Admin token from the ADMIN_TOKEN env var. Used only by the /v1/admin/* endpoints. A compromised regular API key cannot be used to mint more keys."
      }
    }
  },
  "security": [
    {
      "APIKeyAuth": []
    }
  ]
}